Change the logic for finding a spell to cast from while to for loop
Back in the day I could make stuff work, but it wasn't at all pretty or reasonable to make something work that way...
Changed a while (!pSpell) loop that got each memorized spell until it found a valid one and broke out, to then check for readiness to instead be a for loop, that will check readiness of each spell that is memorized until it's checked them all. Will break if we cast a spell.
Changed SpeedRun check to Moving(pSpawn) function.
removed unneeded casting to (int) from spell->ManaCost